requirements.txt
streamlit==1.41.1
matplotlib==3.8.2
numpy==1.26.2
groq==0.13.1
toml==0.10.2
fpdf2==2.7.8
validators==0.22.0
pandas==2.1.4